Externalized search result view to make it reusable
