+++ app/save-and-restore/app/src/main/resources/org/phoebus/applications/saveandrestore/ui/search/SearchResultTableView.fxml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<HBox xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.phoebus.applications.saveandrestore.ui.search.SearchResultTableViewController">
+    <TableView fx:id="resultTableView" HBox.hgrow="ALWAYS">
+        <columns>
+            <TableColumn fx:id="typeColumn" editable="false" maxWidth="40.0" minWidth="40.0" prefWidth="40.0" style="-fx-alignment: TOP-CENTER;" text="%searchTableEntryTypeColumn" />
+            <TableColumn fx:id="nameColumn" editable="false" prefWidth="200.0" style="-fx-alignment: TOP-LEFT;" text="%searchTableNameColumn" />
+            <TableColumn fx:id="commentColumn" editable="false" prefWidth="250.0" text="%searchTableCommentColumn" />
+            <TableColumn fx:id="tagsColumn" editable="false" prefWidth="60.0" text="%searchTableTagsColumn" />
+            <TableColumn fx:id="createdColumn" editable="false" prefWidth="160.0" style="-fx-alignment: TOP-LEFT;" styleClass="timestamp-column" text="%searchCreatedColumn" />
+            <TableColumn fx:id="lastUpdatedColumn" editable="false" prefWidth="160.0" style="-fx-alignment: TOP-LEFT;" styleClass="timestamp-column" text="%searchLastUpdatedColumn" />
+            <TableColumn fx:id="userColumn" editable="false" prefWidth="100.0" style="-fx-alignment: TOP-LEFT;" text="%searchUserColumn" />
+        </columns>
+        <columnResizePolicy>
+            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
+        </columnResizePolicy>
+    </TableView>
+</HBox>
